update ID type

+++ Hahn.ApplicatonProcess.May2020.Domain/Entities/Base/Entity.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 namespace Hahn.ApplicatonProcess.May2020.Domain.Entities.Base
 {
-    public abstract class Entity : EntityBase<int>
+    public abstract class Entity : EntityBase<long>
     {
     }
 }
